[Refactoring]FileService
파일 등록 인터페이스 추가 및 의존관계 변경
